Minor bug fix: Fixed bug in check of ressource
During disconnect the resource is checked if it exists => is connected. If the connection state is lost, the resource is an integer of '0'. This will yield a positiv check with isset($this->_resource). This bugfix prevents an warning triggered by fclose, due an invalid input (expexted resource or integer) during disconnect.