Disconnect when server returns bad response in an atomic pipeline
Predis' atomic pipelines are not throwing a ServerException when Redis
returns an error.  Instead, the code considers it to be a ClientException
and maintains the connection.

This patch will convert server errors to ServerExceptions and disconnect.
The Client object will re-connect on the next command attempt.

This patch also disconnects from the server if no response or the partial
response does not include a server error, as the connection state appears
to be in a bad state and should not be reused.