Implement command queues as the new basis for pipelining.
The scope of these classes in the new Predis\Pipeline\Queue namespace is
to take instances of Predis\Command\CommandInterface and enqueue them so
that they can be flushed to the desired target connection.

As a general rule, unless otherwise stated, flushing the queue returns a
traversable that MUST always be consumed by the caller to make sure all
pending responses on the target connection are cleared to avoid protocol
desynchronization issues.

Queue strategies implemented in the Predis\Pipeline\Queue namespace are:

- Basic: default command queue that implements a standard queue strategy
  where commands are enqueued until a flush is explicitly requested over
  the target connection.

- FireAndForget: queued commands are sent over the target connection but
  responses are not read back from the server. This is accomplished by
  closing the underlying connection in order to quickly drop any pending
  response, meaning that the returned traversable can be safely ignored.

- Atomic: wraps an existing command queue instance in a MULTI/EXEC block
  so that the whole pipeline is flushed in a transaction hiding all the
  required logic. Due to the very nature of MULTI/EXEC, the traversable
  instance from the inner queue is automatically consumed so only actual
  responses in EXEC response payload are returned to the caller.