Rework pipeline abstraction on top of command queues.
There are almost no changes from a consumer point of view when pipelines
are initialized using Predis\Client::pipeline(): their default behaviour
is the same of Predis v1.1 and the same pipeline options are supported.

This is just on the surface though because now we have just one pipeline
class that internally relies on different command queues to encapsulate
different behaviours (basic pipelining, atomic, fire-and-forget).

Thanks to the fact that command queues always return a Traversable for
responses, now a pipeline can be configured to "stream" replies to the
caller instead of just returning an array. Please note that actually not
all command queues support streaming in a real sense, it mostly depends
on their underlying implementation. Anyway, traversable responses should
be handled with care in order to avoid any desynchronization issue. Also
note that partial flushes are unsupported when a pipeline is configured
to return a traversable response. Pipelines can be configured to return
traversable responses from Predis\Client::pipeline() simply by setting
the new "traversable" option to TRUE in the pipeline options dictionary
(default is FALSE).

Atomic pipelines rely on the atomic command queue, a wrapper for other
command queues, meaning that fire-and-forget pipelines can now be sent
in a MULTI / EXEC context simply by setting both "fire-and-forget" and
"atomic" to true in the pipeline options array.

Furthermore, now pipelines can even support iterable multibulk responses
when using Predis\Connection\CompositeStreamConnection. This is kind of
an obscure feature but when combined with pipelines configured to return
traversable responses it allows to effectively stream *every* element of
the response at the protocol level. Atomic pipelines are supported, too!
And when the pipeline is configured to return plain array responses, any
iterable multibulk response is consumed automatically by the pipeline
and returned as a plain array element. This feature should be used with
even more care than normal traversable responses returned by pipelines.

NOTE: we had to apply minor changes to method signatures in the shared
Predis\ClientContextInterface interface so with this commit we adapted
Predis\Transaction\MultiExec to these changes.