Use Replicas For Read
* Added separate pool and slot map for cluster replicas
* Use replicas to serve read requests

This adds a dependency on the replication strategy for detecting read commands. The detection could be moved to the redis cluster strategy or to a separate class included in both strategies.

Also this doesn't allow for the user to configure if and how the replicas should be used. Future improvements could be:
* prefer replicas for reads, fallback to primary
* randomly select between replicas and primaries to evenly distribute

This splits up the slots of the primary between replicas. Another idea would be to share the slots between all replicas and select the replica (randomly) for each command